The Twin Prime conjecture
The Twin Prime conjecture states that there are infinitely many twin prime pairs (two prime numbers that are two apart).

One way to try to prove the twin prime conjecture is to find a larger and larger prime gap between twin prime pairs. If we can find a gap that is larger than any other gap that we know of, then we can conjecture that there are infinitely many twin prime pairs.

We can start by creating a list of prime numbers. We can do this by starting with a list of all integers, and then removing all integers that are not prime.

Next, we can create a list of all twin prime pairs. To do this, we can iterate through the list of prime numbers, and for each prime number, we can check if the next prime number is two away. If it is, we can add the pair to our list of twin prime pairs.

Finally, we can iterate through the list of twin prime pairs, and keep track of the largest gap between a pair. The largest gap that we find will be an upper bound on the twin prime conjecture.